+++ lib/nostr_sdk.dart
@@ -10,6 +10,7 @@ export 'src/rust/api/protocol/event/unsigned.dart';
 export 'src/rust/api/protocol/key.dart';
 export 'src/rust/api/protocol/key/public_key.dart';
 export 'src/rust/api/protocol/key/secret_key.dart';
+export 'src/rust/api/protocol/types/filter.dart';
 export 'src/rust/api/protocol/signer.dart';
 export 'src/rust/api/relay/options.dart';
 export 'src/rust/api/client.dart';